requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.2.1
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fake-useragent==2.0.3
filelock==3.18.0
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.5.0
fsspec==2025.5.1
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
litellm==1.72.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
numpy==2.2.1
openai==1.85.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.23.0
